Resume match 1/requirements.txt
Flask>=3.0.0
numpy>=1.26.0
PyMuPDF>=1.23.5
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
supabase>=2.0.0
python-docx>=1.1.0
pandas>=2.1.1